+++ lib/easypost/services.rb
@@ -28,7 +28,7 @@ require_relative 'services/refund'
 require_relative 'services/report'
 require_relative 'services/scan_form'
 require_relative 'services/shipment'
-require_relative 'services/smartrate'
+require_relative 'services/smart_rate'
 require_relative 'services/tracker'
 require_relative 'services/user'
 require_relative 'services/webhook'
