+++ lib/easypost/services/smart_rate.rb
